move more physical expr
