+++ datafusion-physical-expr/Cargo.toml
@@ -33,6 +33,10 @@ name = "datafusion_physical_expr"
 path = "src/lib.rs"
 
 [features]
+default = ["crypto_expressions", "regex_expressions", "unicode_expressions"]
+crypto_expressions = ["md-5", "sha2", "blake2", "blake3"]
+regex_expressions = ["regex"]
+unicode_expressions = ["unicode-segmentation"]
 
 [dependencies]
 datafusion-common = { path = "../datafusion-common", version = "7.0.0" }
@@ -41,3 +45,13 @@ arrow = { version = "9.0.0", features = ["prettyprint"] }
 paste = "^1.0"
 ahash = { version = "0.7", default-features = false }
 ordered-float = "2.10"
+lazy_static = { version = "^1.4.0" }
+md-5 = { version = "^0.10.0", optional = true }
+sha2 = { version = "^0.10.1", optional = true }
+blake2 = { version = "^0.10.2", optional = true }
+blake3 = { version = "1.0", optional = true }
+rand = "0.8"
+hashbrown = { version = "0.12", features = ["raw"] }
+chrono = { version = "0.4", default-features = false }
+regex = { version = "^1.4.3", optional = true }
+unicode-segmentation = { version = "^1.7.1", optional = true }
