+++ datafusion/Cargo.toml
@@ -40,9 +40,9 @@ path = "src/lib.rs"
 [features]
 default = ["crypto_expressions", "regex_expressions", "unicode_expressions"]
 simd = ["arrow/simd"]
-crypto_expressions = ["md-5", "sha2", "blake2", "blake3"]
-regex_expressions = ["regex"]
-unicode_expressions = ["unicode-segmentation"]
+crypto_expressions = [ "datafusion-physical-expr/crypto_expressions" ]
+unicode_expressions = ["datafusion-physical-expr/regex_expressions"]
+regex_expressions = ["datafusion-physical-expr/regex_expressions"]
 pyarrow = ["pyo3", "arrow/pyarrow", "datafusion-common/pyarrow"]
 # Used for testing ONLY: causes all values to hash to the same value (test for collisions)
 force_hash_collisions = []
@@ -72,13 +72,7 @@ pin-project-lite= "^0.2.7"
 tokio = { version = "1.0", features = ["macros", "rt", "rt-multi-thread", "sync", "fs", "parking_lot"] }
 tokio-stream = "0.1"
 log = "^0.4"
-md-5 = { version = "^0.10.0", optional = true }
-sha2 = { version = "^0.10.1", optional = true }
-blake2 = { version = "^0.10.2", optional = true }
-blake3 = { version = "1.0", optional = true }
 ordered-float = "2.10"
-unicode-segmentation = { version = "^1.7.1", optional = true }
-regex = { version = "^1.4.3", optional = true }
 lazy_static = { version = "^1.4.0" }
 smallvec = { version = "1.6", features = ["union"] }
 rand = "0.8"
